Moving sdk lib files around...
