+++ src/assets/scss/_vars.scss
@@ -13,4 +13,4 @@ $color-the-orange: #ff7e00;
 
 $breakpoint-tablet: 720px;
 
-@import "../../components/SDK/SDK-globals";
+@import "../../sdk/SDK-globals";
