+++ src/sdk/Layouts/LayoutVertical.jsx
