+++ src/sdk/VerticalNav/VerticalNav.scss
